[auto-id] change the data type from long to int

+++ src/main/java/com/recipes/DTO/Recipe.java
@@ -5,10 +5,10 @@ import java.util.List;
 
 public class Recipe {
 
-    private long id;
+    private int id;
     private List<Ingredient> ingredients;
     private String howElaborate;
-    private long userId;
+    private int userId;
 
     public Recipe() {
         ingredients = null;
@@ -20,11 +20,11 @@ public class Recipe {
         this.howElaborate = howElaborate;
     }
 
-    public long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(long id) {
+    public void setId(int id) {
         this.id = id;
     }
 
@@ -44,11 +44,11 @@ public class Recipe {
         this.ingredients = ingredients;
     }
 
-    public long getUserId() {
+    public int getUserId() {
         return userId;
     }
 
-    public void setUserId(long userId) {
+    public void setUserId(int userId) {
         this.userId = userId;
     }
 
@@ -57,7 +57,7 @@ public class Recipe {
                 ingredients.size() > 0;
     }
 
-    public boolean hasId(long id) {
+    public boolean hasId(int id) {
         return this.id == id;
     }
 
@@ -70,7 +70,7 @@ public class Recipe {
         }
     }
 
-    public boolean isOwner(long id) {
+    public boolean isOwner(int id) {
         return this.userId == id;
     }
