+++ src/main/java/com/recipes/Services/IUserServices.java
@@ -10,7 +10,7 @@ public interface IUserServices {
 
     public List<User> getUserList();
 
-    public User findUserbyId(long id);
+    public User findUserbyId(int id);
 
-    public User updateUserInfo(long id, User dataToUpdate, int userId);
+    public User updateUserInfo(int id, User dataToUpdate, int userId);
 }
